ACME: improved handling of duplicate and wildcard domains.
Previously, all domains specified in the server_name directives were
added to the CSR "as is".  In most cases this did not cause any
problems, but it could lead to ACME server errors if the domain list
contained both wildcard domains and matching non-wildcard domains.
Now non-wildcard domains are excluded from the list if a matching
wildcard domain already exists, and duplicate domains are also
removed.