ACME: fixed ACME client disabling issue.
Previously, if an ACME client had been referenced in the "stream" block
but not in the "http" block, it was disabled during configuration
parsing, issued an "ACME client is not defined but referenced" message,
and never fetched a certificate. This has been fixed.

The patch also removes the requirement to place the "http" block
before the "stream" block.
(grafted from cc6483968f88bcab86a5874f854d9a63e83bd7a3)

--HG--
branch : 1.10